+++ CHANGELOG.md
@@ -6,12 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- paging: new package implementing [XEP-0059: Result Set Management]
 - stanza: new functions `AddID` and `AddOriginID` to support unique and stable
   stanza IDs
 - version: new package implementing [XEP-0092: Software Version]
 - xmpp: new `UnmarshalIQ`, `UnmarshalIQElement`, `IterIQ`, and `IterIQElement`
   methods
 
+[XEP-0059: Result Set Management]: https://xmpp.org/extensions/xep-0059.html
 [XEP-0092: Software Version]: https://xmpp.org/extensions/xep-0092.html